increase coverage and clean up
