+++ e2e-test/ui/playwright/pages/stats-v2/stats-tab.page.ts
@@ -36,20 +36,23 @@ export class StatsTabPage extends BasePage {
   readonly queryCountChart: Locator;
   readonly storageSizeCard: Locator;
   readonly storageSizeChart: Locator;
-  readonly changeHistoryCard: Locator;
   readonly changeHistoryChart: Locator;
 
   // Column Stats
   readonly columnStatsContainer: Locator;
-  readonly columnStatsSearchBar: Locator;
 
   // Time Range Selectors (chart-specific to avoid strict mode violations)
   readonly rowCountTimeRangeSelect: Locator;
-  readonly queryCountTimeRangeSelect: Locator;
-  readonly storageSizeTimeRangeSelect: Locator;
 
-  // Common
-  readonly loadingSpinner: Locator;
+  // Chart Empty States
+  readonly rowCountChartEmpty: Locator;
+  readonly queryCountChartEmpty: Locator;
+  readonly storageSizeChartEmpty: Locator;
+  readonly changeHistoryChartEmpty: Locator;
+
+  // Changes Card & Schema Drawer
+  readonly changesCard: Locator;
+  readonly schemaFieldDrawerContent: Locator;
 
   constructor(page: Page, logger?: DataHubLogger, logDir?: string) {
     super(page, logger, logDir);
@@ -69,20 +72,46 @@ export class StatsTabPage extends BasePage {
     this.queryCountChart = page.getByTestId('query-count-chart');
     this.storageSizeCard = page.getByTestId('storage-size-card');
     this.storageSizeChart = page.getByTestId('storage-size-chart');
-    this.changeHistoryCard = page.getByTestId('change-history-card');
     this.changeHistoryChart = page.getByTestId('change-history-chart');
 
     // Column Stats
     this.columnStatsContainer = page.getByTestId('column-stats-container');
-    this.columnStatsSearchBar = page.getByTestId('column-stats-search-bar');
 
     // Time Range Selectors - scoped to specific chart cards
     this.rowCountTimeRangeSelect = this.rowCountCard.getByTestId('timerange-select');
-    this.queryCountTimeRangeSelect = this.queryCountCard.getByTestId('timerange-select');
-    this.storageSizeTimeRangeSelect = this.storageSizeCard.getByTestId('timerange-select');
 
-    // Common
-    this.loadingSpinner = page.getByRole('status');
+    // Chart Empty States
+    this.rowCountChartEmpty = page.getByTestId('row-count-chart-empty');
+    this.queryCountChartEmpty = page.getByTestId('query-count-chart-empty');
+    this.storageSizeChartEmpty = page.getByTestId('storage-size-chart-empty');
+    this.changeHistoryChartEmpty = page.getByTestId('change-history-chart-empty');
+
+    // Changes Card & Schema Drawer
+    this.changesCard = page.getByTestId('changes-card');
+    this.schemaFieldDrawerContent = page.getByTestId('schema-field-drawer-content');
+  }
+
+  // ── Dynamic Selectors (Parameterized) ─────────────────────────────────────
+
+  /**
+   * Get a time range dropdown option by name
+   */
+  private getTimeRangeOption(timeRange: string): Locator {
+    return this.page.getByTestId(`option-${timeRange}`);
+  }
+
+  /**
+   * Get a time range selected value indicator
+   */
+  private getTimeRangeSelectedValue(timeRange: string): Locator {
+    return this.page.getByTestId(`value-${timeRange}`);
+  }
+
+  /**
+   * Get a column row in the statistics table by column name
+   */
+  private getColumnRow(columnName: string): Locator {
+    return this.columnStatsContainer.getByTestId(`row-${columnName}`);
   }
 
   // ── Navigation ────────────────────────────────────────────────────────────
@@ -99,19 +128,11 @@ export class StatsTabPage extends BasePage {
   // ── Tab Visibility ────────────────────────────────────────────────────────
 
   /**
-   * Verify stats tab exists and is visible
+   * Get stats tab element via role (for attribute assertions)
+   * Returns the tab element itself rather than testid-based locator
    */
-  async verifyStatsTabIsVisible(): Promise<void> {
-    this.logger?.step('verifyStatsTabIsVisible');
-    await this.statsTab.waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
-  }
-
-  /**
-   * Verify stats container is loaded and visible
-   */
-  async verifyStatsContainerLoaded(): Promise<void> {
-    this.logger?.step('verifyStatsContainerLoaded');
-    await this.statsContainer.waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
+  getStatsTabElement(): Locator {
+    return this.page.getByRole('tab', { name: 'Stats' });
   }
 
   /**
@@ -144,14 +165,6 @@ export class StatsTabPage extends BasePage {
     this.logger?.step('verifyLastMonthStatsCard', { cardTestId, expectedValue });
     const card = this.lastMonthStatsContainer.getByTestId(cardTestId);
     await expect(card).toContainText(expectedValue);
-  }
-
-  /**
-   * Get value from a stat card
-   */
-  async getStatsCardValue(container: Locator, cardTestId: string): Promise<string> {
-    const card = container.getByTestId(cardTestId);
-    return card.textContent().then((content) => content?.trim() || '');
   }
 
   /**
@@ -205,8 +218,7 @@ export class StatsTabPage extends BasePage {
    */
   async verifyChangesCardVisible(): Promise<void> {
     this.logger?.step('verifyChangesCardVisible');
-    const changesCard = this.page.getByTestId('changes-card');
-    await expect(changesCard).toBeVisible({ timeout: TIMEOUTS.MEDIUM });
+    await expect(this.changesCard).toBeVisible({ timeout: TIMEOUTS.MEDIUM });
   }
 
   /**
@@ -214,8 +226,7 @@ export class StatsTabPage extends BasePage {
    */
   async verifyChangesCardValue(expectedValue: string): Promise<void> {
     this.logger?.step('verifyChangesCardValue', { expectedValue });
-    const changesCard = this.page.getByTestId('changes-card');
-    await expect(changesCard).toContainText(expectedValue);
+    await expect(this.changesCard).toContainText(expectedValue);
   }
 
   /**
@@ -223,7 +234,7 @@ export class StatsTabPage extends BasePage {
    */
   async verifyChangesCardButtonVisible(): Promise<void> {
     this.logger?.step('verifyChangesCardButtonVisible');
-    const button = this.page.getByTestId('changes-card').getByRole('button');
+    const button = this.changesCard.getByRole('button');
     await expect(button).toBeVisible({ timeout: TIMEOUTS.MEDIUM });
   }
 
@@ -242,8 +253,7 @@ export class StatsTabPage extends BasePage {
    */
   async verifyRowCountChartIsEmpty(): Promise<void> {
     this.logger?.step('verifyRowCountChartIsEmpty');
-    const emptyState = this.page.getByTestId('row-count-chart-empty');
-    await emptyState.waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
+    await this.rowCountChartEmpty.waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
   }
 
   // ── Charts - Query Count ──────────────────────────────────────────────────
@@ -261,8 +271,7 @@ export class StatsTabPage extends BasePage {
    */
   async verifyQueryCountChartIsEmpty(): Promise<void> {
     this.logger?.step('verifyQueryCountChartIsEmpty');
-    const emptyState = this.page.getByTestId('query-count-chart-empty');
-    await emptyState.waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
+    await this.queryCountChartEmpty.waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
   }
 
   // ── Charts - Storage Size ─────────────────────────────────────────────────
@@ -280,8 +289,7 @@ export class StatsTabPage extends BasePage {
    */
   async verifyStorageSizeChartIsEmpty(): Promise<void> {
     this.logger?.step('verifyStorageSizeChartIsEmpty');
-    const emptyState = this.page.getByTestId('storage-size-chart-empty');
-    await emptyState.waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
+    await this.storageSizeChartEmpty.waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
   }
 
   // ── Charts - Change History ───────────────────────────────────────────────
@@ -299,23 +307,10 @@ export class StatsTabPage extends BasePage {
    */
   async verifyChangeHistoryChartIsEmpty(): Promise<void> {
     this.logger?.step('verifyChangeHistoryChartIsEmpty');
-    const emptyState = this.page.getByTestId('change-history-chart-empty');
-    await emptyState.waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
+    await this.changeHistoryChartEmpty.waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
   }
 
   // ── Time Range Filtering ──────────────────────────────────────────────────
-
-  /**
-   * Select a time range from the dropdown (e.g., MONTH, WEEK, ALL)
-   * Uses row count chart selector as representative (all charts have same options)
-   */
-  async selectTimeRange(timeRange: string): Promise<void> {
-    this.logger?.step('selectTimeRange', { timeRange });
-    await this.rowCountTimeRangeSelect.click({ timeout: TIMEOUTS.MEDIUM });
-    await this.page.getByTestId(`option-${timeRange}`).click({ timeout: TIMEOUTS.SHORT });
-    // Wait for charts to update
-    await this.page.waitForLoadState('networkidle');
-  }
 
   /**
    * Verify time range selector is visible
@@ -338,6 +333,8 @@ export class StatsTabPage extends BasePage {
     }
   }
 
+  // ── Time Range Verification ────────────────────────────────────────────────
+
   /**
    * Verify a specific time range option is available
    * Uses row count chart selector to open dropdown
@@ -347,8 +344,7 @@ export class StatsTabPage extends BasePage {
     // Click dropdown to open it
     await this.rowCountTimeRangeSelect.click();
     // Wait for the option to become visible
-    const option = this.page.getByTestId(`option-${timeRange}`);
-    await option.waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
+    await this.getTimeRangeOption(timeRange).waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
   }
 
   /**
@@ -356,8 +352,7 @@ export class StatsTabPage extends BasePage {
    */
   async verifyTimeRangeOptionDoesNotExist(timeRange: string): Promise<void> {
     this.logger?.step('verifyTimeRangeOptionDoesNotExist', { timeRange });
-    const option = this.page.getByTestId(`option-${timeRange}`);
-    const count = await option.count();
+    const count = await this.getTimeRangeOption(timeRange).count();
     if (count > 0) {
       throw new Error(`Time range option "${timeRange}" should not exist but was found`);
     }
@@ -368,78 +363,18 @@ export class StatsTabPage extends BasePage {
    */
   async verifyTimeRangeSelected(timeRange: string): Promise<void> {
     this.logger?.step('verifyTimeRangeSelected', { timeRange });
-    const selectedValue = this.page.getByTestId(`value-${timeRange}`);
-    await expect(selectedValue).toBeVisible();
+    await expect(this.getTimeRangeSelectedValue(timeRange)).toBeVisible();
   }
 
   // ── Column Stats Table ────────────────────────────────────────────────────
-
-  /**
-   * Verify column stats table is visible
-   */
-  async verifyColumnStatsTableIsVisible(): Promise<void> {
-    this.logger?.step('verifyColumnStatsTableIsVisible');
-    await this.columnStatsContainer.waitFor({
-      state: 'visible',
-      timeout: TIMEOUTS.MEDIUM,
-    });
-  }
-
-  /**
-   * Get count of column rows in table
-   */
-  async getColumnRowCount(): Promise<number> {
-    this.logger?.step('getColumnRowCount');
-    return this.columnStatsContainer.getByTestId(/^row-/).count();
-  }
-
-  /**
-   * Verify a column exists in the stats table
-   */
-  async verifyColumnExists(columnName: string): Promise<void> {
-    this.logger?.step('verifyColumnExists', { columnName });
-    const row = this.columnStatsContainer.getByTestId(`row-${columnName}`);
-    await row.waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
-  }
-
-  /**
-   * Verify a column does NOT exist in the stats table
-   */
-  async verifyColumnDoesNotExist(columnName: string): Promise<void> {
-    this.logger?.step('verifyColumnDoesNotExist', { columnName });
-    const row = this.columnStatsContainer.getByTestId(`row-${columnName}`);
-    const count = await row.count();
-    if (count > 0) {
-      throw new Error(`Column "${columnName}" should not exist but was found`);
-    }
-  }
 
   /**
    * Click on a column row to open details
    */
   async clickColumnRow(columnName: string): Promise<void> {
     this.logger?.step('clickColumnRow', { columnName });
-    const row = this.columnStatsContainer.getByTestId(`row-${columnName}`);
-    await row.click({ timeout: TIMEOUTS.SHORT });
-    await this.page.getByTestId('schema-field-drawer-content').waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
-  }
-
-  /**
-   * Search/filter columns by name
-   */
-  async searchColumns(searchTerm: string): Promise<void> {
-    this.logger?.step('searchColumns', { searchTerm });
-    await this.columnStatsSearchBar.fill(searchTerm);
-    await this.page.waitForLoadState('networkidle');
-  }
-
-  /**
-   * Clear column search filter
-   */
-  async clearColumnSearch(): Promise<void> {
-    this.logger?.step('clearColumnSearch');
-    await this.columnStatsSearchBar.clear();
-    await this.page.waitForLoadState('networkidle');
+    await this.getColumnRow(columnName).click({ timeout: TIMEOUTS.SHORT });
+    await this.schemaFieldDrawerContent.waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
   }
 
   /**
@@ -448,7 +383,7 @@ export class StatsTabPage extends BasePage {
    */
   async verifyColumnFields(columnName: string, fields: Record<string, string>): Promise<void> {
     this.logger?.step('verifyColumnFields', { columnName, fields });
-    const row = this.columnStatsContainer.getByTestId(`row-${columnName}`);
+    const row = this.getColumnRow(columnName);
     await row.evaluate((el) => el.scrollIntoView());
 
     for (const [fieldTestId, expectedValue] of Object.entries(fields)) {
@@ -462,7 +397,7 @@ export class StatsTabPage extends BasePage {
    */
   async verifySchemaFieldDrawerOpen(): Promise<void> {
     this.logger?.step('verifySchemaFieldDrawerOpen');
-    const drawerContent = this.page.getByTestId('schema-field-drawer-content');
+    const drawerContent = this.schemaFieldDrawerContent;
     await drawerContent.waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
   }
