+++ e2e-test/ui/playwright/tests/stats-v2/v2-charts.spec.ts
@@ -15,17 +15,18 @@
 
 import { test } from '../../fixtures/base-test';
 import { StatsTabPage } from '../../pages/stats-v2/stats-tab.page';
-import { getSampleProfile, getSampleUsageStats } from '../../helpers/stats-mock-helper';
+import {
+  getSampleProfile,
+  getSampleUsageStats,
+  setupChartsData,
+  setupChartsDataEmpty,
+  setupChartsDataNoPermissions,
+} from '../../factories/mock-responses/stats';
+import { TEST_DATASET_URN, TIME_RANGES } from './stats-constants';
 
 const TEST_DATA = {
-  DATASET_URN: 'urn:li:dataset:(urn:li:dataPlatform:postgres,playwright_stats_test,PROD)',
-  TIME_RANGES: {
-    WEEK: 'WEEK',
-    MONTH: 'MONTH',
-    QUARTER: 'QUARTER',
-    HALF_OF_YEAR: 'HALF_OF_YEAR',
-    YEAR: 'YEAR',
-  },
+  DATASET_URN: TEST_DATASET_URN,
+  TIME_RANGES,
 } as const;
 
 test.use({ featureName: 'stats-v2' });
@@ -35,62 +36,11 @@ test.describe('Statistics Charts', () => {
 
   test.beforeEach(async ({ page, logger, logDir }) => {
     statsPage = new StatsTabPage(page, logger, logDir);
-    // Feature flag is already enabled by default in test environment
   });
 
   test('should be available when there are some data', async ({ apiMock }) => {
     const timestamp = Math.floor(Date.now() / 1000) * 1000;
-    const sampleProfile = getSampleProfile(timestamp);
-    const sampleUsageStats = getSampleUsageStats(timestamp);
-
-    // Mock getDataset with latestFullTableProfile to enable stats tab
-    await apiMock.mockGraphQL('getDataset', {
-      dataset: {
-        __typename: 'Dataset',
-        urn: TEST_DATA.DATASET_URN,
-        latestFullTableProfile: [sampleProfile],
-        latestPartitionProfile: [],
-        privileges: {
-          __typename: 'DatasetPrivileges',
-          canViewDatasetProfile: true,
-          canViewDatasetUsage: true,
-          canViewDatasetOperations: true,
-          canEditDatasetProperties: true,
-        },
-      },
-    });
-
-    // Mock detail endpoints that provide the actual data
-    await apiMock.mockGraphQL('getDataProfiles', {
-      dataset: {
-        __typename: 'Dataset',
-        datasetProfiles: [sampleProfile],
-      },
-    });
-
-    await apiMock.mockGraphQL('getTimeRangeUsageAggregations', {
-      dataset: {
-        __typename: 'Dataset',
-        usageStats: {
-          __typename: 'UsageAggregation',
-          buckets: sampleUsageStats.buckets,
-        },
-      },
-    });
-
-    await apiMock.mockGraphQL('getDatasetTimeseriesCapability', {
-      dataset: {
-        __typename: 'Dataset',
-        timeseriesCapabilities: {
-          __typename: 'TimeseriesCapabilities',
-          assetStats: {
-            __typename: 'AssetStats',
-            oldestDatasetProfileTime: timestamp,
-            oldestDatasetUsageTime: timestamp,
-          },
-        },
-      },
-    });
+    await setupChartsData(apiMock, timestamp, TEST_DATA.DATASET_URN);
 
     await statsPage.navigateToDatasetStats(TEST_DATA.DATASET_URN);
 
@@ -101,57 +51,7 @@ test.describe('Statistics Charts', () => {
   });
 
   test('should be empty when there are no any data', async ({ apiMock }) => {
-    const timestamp = Math.floor(Date.now() / 1000) * 1000;
-    const sampleProfile = getSampleProfile(timestamp);
-
-    // Mock getDataset with profile to enable stats tab
-    await apiMock.mockGraphQL('getDataset', {
-      dataset: {
-        __typename: 'Dataset',
-        urn: TEST_DATA.DATASET_URN,
-        latestFullTableProfile: [sampleProfile],
-        latestPartitionProfile: [],
-        privileges: {
-          __typename: 'DatasetPrivileges',
-          canViewDatasetProfile: true,
-          canViewDatasetUsage: true,
-          canViewDatasetOperations: true,
-          canEditDatasetProperties: true,
-        },
-      },
-    });
-
-    // Mock detail endpoints with empty data
-    await apiMock.mockGraphQL('getDataProfiles', {
-      dataset: {
-        __typename: 'Dataset',
-        datasetProfiles: [],
-      },
-    });
-
-    await apiMock.mockGraphQL('getTimeRangeUsageAggregations', {
-      dataset: {
-        __typename: 'Dataset',
-        usageStats: {
-          __typename: 'UsageAggregation',
-          buckets: [],
-        },
-      },
-    });
-
-    await apiMock.mockGraphQL('getDatasetTimeseriesCapability', {
-      dataset: {
-        __typename: 'Dataset',
-        timeseriesCapabilities: {
-          __typename: 'TimeseriesCapabilities',
-          assetStats: {
-            __typename: 'AssetStats',
-            oldestDatasetProfileTime: null,
-            oldestDatasetUsageTime: null,
-          },
-        },
-      },
-    });
+    await setupChartsDataEmpty(apiMock, TEST_DATA.DATASET_URN);
 
     await statsPage.navigateToDatasetStats(TEST_DATA.DATASET_URN);
 
@@ -162,58 +62,7 @@ test.describe('Statistics Charts', () => {
   });
 
   test('should hide time filter when there are no data', async ({ apiMock }) => {
-    const timestamp = Math.floor(Date.now() / 1000) * 1000;
-    const sampleProfile = getSampleProfile(timestamp);
-
-    // Mock getDataset with profile to enable stats tab
-    await apiMock.mockGraphQL('getDataset', {
-      dataset: {
-        __typename: 'Dataset',
-        urn: TEST_DATA.DATASET_URN,
-        latestFullTableProfile: [sampleProfile],
-        latestPartitionProfile: [],
-        privileges: {
-          __typename: 'DatasetPrivileges',
-          canViewDatasetProfile: true,
-          canViewDatasetUsage: true,
-          canViewDatasetOperations: true,
-          canEditDatasetProperties: true,
-        },
-      },
-    });
-
-    // Mock detail endpoints with empty data
-    await apiMock.mockGraphQL('getDataProfiles', {
-      dataset: {
-        __typename: 'Dataset',
-        datasetProfiles: [],
-      },
-    });
-
-    await apiMock.mockGraphQL('getTimeRangeUsageAggregations', {
-      dataset: {
-        __typename: 'Dataset',
-        usageStats: {
-          __typename: 'UsageAggregation',
-          buckets: [],
-        },
-      },
-    });
-
-    await apiMock.mockGraphQL('getDatasetTimeseriesCapability', {
-      dataset: {
-        __typename: 'Dataset',
-        timeseriesCapabilities: {
-          __typename: 'TimeseriesCapabilities',
-          assetStats: {
-            __typename: 'AssetStats',
-            oldestDatasetProfileTime: null,
-            oldestDatasetUsageTime: null,
-          },
-        },
-      },
-    });
-
+    await setupChartsDataEmpty(apiMock, TEST_DATA.DATASET_URN);
     await statsPage.navigateToDatasetStats(TEST_DATA.DATASET_URN);
 
     // Time range selector should not exist when no data
@@ -351,6 +200,74 @@ test.describe('Statistics Charts', () => {
     await statsPage.verifyTimeRangeOptionExists(TEST_DATA.TIME_RANGES.YEAR);
   });
 
+  test('should show time filter with expected options when more than 6 months of data available', async ({
+    apiMock,
+  }) => {
+    const now = Date.now();
+    const sixMonthsAgo = now - (6 * 30 + 7) * 24 * 60 * 60 * 1000;
+    const timestamp = Math.floor(sixMonthsAgo / 1000) * 1000;
+
+    const sampleProfile = getSampleProfile(timestamp);
+    const sampleUsageStats = getSampleUsageStats(timestamp);
+
+    await apiMock.mockGraphQL('getDataset', {
+      dataset: {
+        __typename: 'Dataset',
+        urn: TEST_DATA.DATASET_URN,
+        latestFullTableProfile: [sampleProfile],
+        latestPartitionProfile: [],
+        privileges: {
+          __typename: 'DatasetPrivileges',
+          canViewDatasetProfile: true,
+          canViewDatasetUsage: true,
+          canViewDatasetOperations: true,
+          canEditDatasetProperties: true,
+        },
+      },
+    });
+
+    await apiMock.mockGraphQL('getDataProfiles', {
+      dataset: {
+        __typename: 'Dataset',
+        datasetProfiles: [sampleProfile],
+      },
+    });
+
+    await apiMock.mockGraphQL('getTimeRangeUsageAggregations', {
+      dataset: {
+        __typename: 'Dataset',
+        usageStats: {
+          __typename: 'UsageAggregation',
+          buckets: sampleUsageStats.buckets,
+        },
+      },
+    });
+
+    await apiMock.mockGraphQL('getDatasetTimeseriesCapability', {
+      dataset: {
+        __typename: 'Dataset',
+        timeseriesCapabilities: {
+          __typename: 'TimeseriesCapabilities',
+          assetStats: {
+            __typename: 'AssetStats',
+            oldestDatasetProfileTime: timestamp,
+            oldestDatasetUsageTime: timestamp,
+          },
+        },
+      },
+    });
+
+    await statsPage.navigateToDatasetStats(TEST_DATA.DATASET_URN);
+
+    await statsPage.verifyTimeRangeSelectorExists();
+    await statsPage.verifyTimeRangeSelected(TEST_DATA.TIME_RANGES.MONTH);
+    await statsPage.verifyTimeRangeOptionExists(TEST_DATA.TIME_RANGES.WEEK);
+    await statsPage.verifyTimeRangeOptionExists(TEST_DATA.TIME_RANGES.MONTH);
+    await statsPage.verifyTimeRangeOptionExists(TEST_DATA.TIME_RANGES.QUARTER);
+    await statsPage.verifyTimeRangeOptionExists(TEST_DATA.TIME_RANGES.HALF_OF_YEAR);
+    await statsPage.verifyTimeRangeOptionExists(TEST_DATA.TIME_RANGES.YEAR);
+  });
+
   test('should show time filter with expected options when more than 3 months of data available', async ({
     apiMock,
   }) => {
@@ -431,8 +348,8 @@ test.describe('Statistics Charts', () => {
     apiMock,
   }) => {
     const now = Date.now();
-    const oneFiftyDaysAgo = now - 150 * 24 * 60 * 60 * 1000;
-    const timestamp = Math.floor(oneFiftyDaysAgo / 1000) * 1000;
+    const oneMonthAndOneWeekAgo = now - (30 + 7) * 24 * 60 * 60 * 1000;
+    const timestamp = Math.floor(oneMonthAndOneWeekAgo / 1000) * 1000;
 
     const sampleProfile = getSampleProfile(timestamp);
     const sampleUsageStats = getSampleUsageStats(timestamp);
@@ -497,16 +414,16 @@ test.describe('Statistics Charts', () => {
     await statsPage.verifyTimeRangeOptionExists(TEST_DATA.TIME_RANGES.WEEK);
     await statsPage.verifyTimeRangeOptionExists(TEST_DATA.TIME_RANGES.MONTH);
     await statsPage.verifyTimeRangeOptionExists(TEST_DATA.TIME_RANGES.QUARTER);
-    await statsPage.verifyTimeRangeOptionExists(TEST_DATA.TIME_RANGES.HALF_OF_YEAR);
 
-    // Year should not be available
+    // Half year and year should not be available
+    await statsPage.verifyTimeRangeOptionDoesNotExist(TEST_DATA.TIME_RANGES.HALF_OF_YEAR);
     await statsPage.verifyTimeRangeOptionDoesNotExist(TEST_DATA.TIME_RANGES.YEAR);
   });
 
   test('should show time filter with expected options when more than 1 week of data available', async ({ apiMock }) => {
     const now = Date.now();
-    const oneMonthAgo = now - 30 * 24 * 60 * 60 * 1000;
-    const timestamp = Math.floor(oneMonthAgo / 1000) * 1000;
+    const oneWeekAndOneDayAgo = now - (7 + 1) * 24 * 60 * 60 * 1000;
+    const timestamp = Math.floor(oneWeekAndOneDayAgo / 1000) * 1000;
 
     const sampleProfile = getSampleProfile(timestamp);
     const sampleUsageStats = getSampleUsageStats(timestamp);
@@ -578,25 +495,7 @@ test.describe('Statistics Charts', () => {
   });
 
   test('should not be available when user has no permissions', async ({ apiMock }) => {
-    const timestamp = Math.floor(Date.now() / 1000) * 1000;
-    const sampleProfile = getSampleProfile(timestamp);
-
-    await apiMock.mockGraphQL('getDataset', {
-      dataset: {
-        __typename: 'Dataset',
-        urn: TEST_DATA.DATASET_URN,
-        latestFullTableProfile: [sampleProfile],
-        latestPartitionProfile: [],
-        privileges: {
-          __typename: 'DatasetPrivileges',
-          canViewDatasetProfile: false,
-          canViewDatasetUsage: false,
-          canViewDatasetOperations: false,
-          canEditDatasetProperties: false,
-        },
-      },
-    });
-
+    await setupChartsDataNoPermissions(apiMock, TEST_DATA.DATASET_URN);
     await statsPage.navigateToDatasetStats(TEST_DATA.DATASET_URN);
 
     // All charts should show no permissions message
