+++ e2e-test/ui/playwright/tests/stats-v2/v2-highlights.spec.ts
@@ -12,252 +12,22 @@
 
 import { test } from '../../fixtures/base-test';
 import { StatsTabPage } from '../../pages/stats-v2/stats-tab.page';
-import { getSampleProfile, getEmptyProfile } from '../../helpers/stats-mock-helper';
+import { getEmptyProfile } from '../../factories/mock-responses/stats';
 import type { ApiMocker } from '../../fixtures/mocking.fixture';
+import { TEST_DATASET_URN, DEFAULT_PRIVILEGES, EXPECTED_STATS } from './stats-constants';
+import { setupBasicDataset, setupUsageStats, setupOperationsStats } from '../../factories/mock-responses/stats';
 
 const TEST_DATA = {
-  DATASET_URN: 'urn:li:dataset:(urn:li:dataPlatform:postgres,playwright_stats_test,PROD)',
+  DATASET_URN: TEST_DATASET_URN,
 } as const;
 
 test.use({ featureName: 'stats-v2' });
 
-// Helper to set up mocks inline in each test
 async function setupDataWithStats(apiMock: ApiMocker) {
   const timestamp = Date.now();
-  const sampleProfile = getSampleProfile(timestamp);
-
-  // Mock getDataset with profile to enable stats tab
-  await apiMock.mockGraphQL('getDataset', {
-    dataset: {
-      __typename: 'Dataset',
-      urn: TEST_DATA.DATASET_URN,
-      latestFullTableProfile: [sampleProfile],
-      latestPartitionProfile: [],
-      privileges: {
-        __typename: 'DatasetPrivileges',
-        canViewDatasetProfile: true,
-        canViewDatasetUsage: true,
-        canViewDatasetOperations: true,
-        canEditDatasetProperties: true,
-      },
-    },
-  });
-
-  // Mock detail endpoints with actual data
-  await apiMock.mockGraphQL('getDataProfiles', {
-    dataset: {
-      __typename: 'Dataset',
-      datasetProfiles: [sampleProfile],
-    },
-  });
-
-  await apiMock.mockGraphQL('getDataProfiles', {
-    dataset: {
-      __typename: 'Dataset',
-      datasetProfiles: [sampleProfile],
-    },
-  });
-
-  await apiMock.mockGraphQL('getDatasetTimeseriesCapability', {
-    dataset: {
-      __typename: 'Dataset',
-      timeseriesCapabilities: {
-        __typename: 'TimeseriesCapabilities',
-        assetStats: {
-          __typename: 'AssetStats',
-          oldestDatasetProfileTime: timestamp,
-        },
-      },
-    },
-  });
-
-  await apiMock.mockGraphQL('getTimeRangeUsageAggregations', {
-    dataset: {
-      __typename: 'Dataset',
-      usageStats: {
-        __typename: 'UsageAggregation',
-        buckets: [
-          {
-            bucket: timestamp,
-            metrics: {
-              __typename: 'UsageAggregationMetrics',
-              totalSqlQueries: 25,
-            },
-            __typename: 'UsageAggregation',
-          },
-        ],
-      },
-    },
-  });
-
-  await apiMock.mockGraphQL('getLastMonthUsageAggregations', {
-    dataset: {
-      __typename: 'Dataset',
-      usageStats: {
-        __typename: 'UsageAggregation',
-        aggregations: {
-          __typename: 'UsageAggregationMetrics',
-          totalSqlQueries: 25,
-          uniqueUserCount: 5,
-          users: [
-            {
-              count: 1,
-              userEmail: 'user1@example.com',
-              user: {
-                __typename: 'CorpUser',
-                urn: 'urn:li:corpuser:user1',
-                username: 'user1',
-                type: 'CORP_USER',
-                properties: {
-                  __typename: 'CorpUserProperties',
-                  displayName: 'User One',
-                  firstName: 'User',
-                  lastName: 'One',
-                  fullName: 'User One',
-                },
-                editableProperties: {
-                  __typename: 'CorpUserEditableProperties',
-                  displayName: 'User One',
-                  pictureLink: '',
-                },
-              },
-              __typename: 'UsageAggregationUser',
-            },
-            {
-              count: 1,
-              userEmail: 'user2@example.com',
-              user: {
-                __typename: 'CorpUser',
-                urn: 'urn:li:corpuser:user2',
-                username: 'user2',
-                type: 'CORP_USER',
-                properties: {
-                  __typename: 'CorpUserProperties',
-                  displayName: 'User Two',
-                  firstName: 'User',
-                  lastName: 'Two',
-                  fullName: 'User Two',
-                },
-                editableProperties: {
-                  __typename: 'CorpUserEditableProperties',
-                  displayName: 'User Two',
-                  pictureLink: '',
-                },
-              },
-              __typename: 'UsageAggregationUser',
-            },
-            {
-              count: 1,
-              userEmail: 'user3@example.com',
-              user: {
-                __typename: 'CorpUser',
-                urn: 'urn:li:corpuser:user3',
-                username: 'user3',
-                type: 'CORP_USER',
-                properties: {
-                  __typename: 'CorpUserProperties',
-                  displayName: 'User Three',
-                  firstName: 'User',
-                  lastName: 'Three',
-                  fullName: 'User Three',
-                },
-                editableProperties: {
-                  __typename: 'CorpUserEditableProperties',
-                  displayName: 'User Three',
-                  pictureLink: '',
-                },
-              },
-              __typename: 'UsageAggregationUser',
-            },
-            {
-              count: 1,
-              userEmail: 'user4@example.com',
-              user: {
-                __typename: 'CorpUser',
-                urn: 'urn:li:corpuser:user4',
-                username: 'user4',
-                type: 'CORP_USER',
-                properties: {
-                  __typename: 'CorpUserProperties',
-                  displayName: 'User Four',
-                  firstName: 'User',
-                  lastName: 'Four',
-                  fullName: 'User Four',
-                },
-                editableProperties: {
-                  __typename: 'CorpUserEditableProperties',
-                  displayName: 'User Four',
-                  pictureLink: '',
-                },
-              },
-              __typename: 'UsageAggregationUser',
-            },
-            {
-              count: 1,
-              userEmail: 'user5@example.com',
-              user: {
-                __typename: 'CorpUser',
-                urn: 'urn:li:corpuser:user5',
-                username: 'user5',
-                type: 'CORP_USER',
-                properties: {
-                  __typename: 'CorpUserProperties',
-                  displayName: 'User Five',
-                  firstName: 'User',
-                  lastName: 'Five',
-                  fullName: 'User Five',
-                },
-                editableProperties: {
-                  __typename: 'CorpUserEditableProperties',
-                  displayName: 'User Five',
-                  pictureLink: '',
-                },
-              },
-              __typename: 'UsageAggregationUser',
-            },
-          ],
-          fields: [],
-        },
-      },
-    },
-  });
-
-  await apiMock.mockGraphQL('getOperationsStats', {
-    dataset: {
-      __typename: 'Dataset',
-      operationsStats: {
-        __typename: 'OperationsAggregation',
-        aggregations: {
-          __typename: 'OperationsAggregationMetrics',
-          totalCreates: 1,
-          totalOperations: 1,
-        },
-      },
-    },
-  });
-
-  await apiMock.mockGraphQL('getOperationsStatsBuckets', {
-    dataset: {
-      __typename: 'Dataset',
-      operationsStats: {
-        __typename: 'OperationsAggregation',
-        aggregations: {
-          __typename: 'OperationsAggregationMetrics',
-          totalCreates: 1,
-        },
-        buckets: [
-          {
-            bucket: timestamp,
-            aggregations: {
-              __typename: 'OperationsAggregationMetrics',
-              totalCreates: 1,
-            },
-            __typename: 'OperationsAggregation',
-          },
-        ],
-      },
-    },
-  });
+  await setupBasicDataset(apiMock, timestamp, TEST_DATASET_URN, DEFAULT_PRIVILEGES);
+  await setupUsageStats(apiMock, timestamp);
+  await setupOperationsStats(apiMock, timestamp, false);
 }
 
 test.describe('Highlight Stats Cards', () => {
@@ -269,28 +39,27 @@ test.describe('Highlight Stats Cards', () => {
   });
 
   test('should show values when the data is available', async ({ apiMock }) => {
-    await apiMock.setFeatureFlags({ showStatsTabRedesign: true });
     await setupDataWithStats(apiMock);
 
     await statsPage.navigateToDatasetStats(TEST_DATA.DATASET_URN);
 
     // Latest stats (Rows, Columns)
     await statsPage.verifyLatestStatsVisible();
-    await statsPage.verifyLatestStatsCard('rows-card', '100');
+    await statsPage.verifyLatestStatsCard('rows-card', EXPECTED_STATS.ROWS);
     await statsPage.verifyLatestStatsCardButtonVisible('rows-card');
-    await statsPage.verifyLatestStatsCard('columns-card', '7');
+    await statsPage.verifyLatestStatsCard('columns-card', EXPECTED_STATS.COLUMNS);
     await statsPage.verifyLatestStatsCardButtonVisible('columns-card');
 
     // Last month stats (Users, Queries)
     await statsPage.verifyLastMonthStatsVisible();
-    await statsPage.verifyLastMonthStatsCard('users-card', '5');
+    await statsPage.verifyLastMonthStatsCard('users-card', EXPECTED_STATS.USERS);
     await statsPage.verifyLastMonthStatsCardButtonVisible('users-card');
-    await statsPage.verifyLastMonthStatsCard('queries-card', '25');
+    await statsPage.verifyLastMonthStatsCard('queries-card', EXPECTED_STATS.QUERIES);
     await statsPage.verifyLastMonthStatsCardButtonVisible('queries-card');
 
     // Changes card
     await statsPage.verifyChangesCardVisible();
-    await statsPage.verifyChangesCardValue('1');
+    await statsPage.verifyChangesCardValue(EXPECTED_STATS.CHANGES);
     await statsPage.verifyChangesCardButtonVisible();
   });
 
@@ -304,13 +73,7 @@ test.describe('Highlight Stats Cards', () => {
         urn: TEST_DATA.DATASET_URN,
         latestFullTableProfile: [emptyProfile],
         latestPartitionProfile: [],
-        privileges: {
-          __typename: 'DatasetPrivileges',
-          canViewDatasetProfile: true,
-          canViewDatasetUsage: true,
-          canViewDatasetOperations: true,
-          canEditDatasetProperties: true,
-        },
+        privileges: DEFAULT_PRIVILEGES,
       },
     });
 
@@ -397,7 +160,7 @@ test.describe('Highlight Stats Cards', () => {
 
     // Rows card - should not show value when profile is empty
     // The card shows "No Data" because rowCount is null
-    await statsPage.verifyLatestStatsCardDoesNotContain('rows-card', '100');
+    await statsPage.verifyLatestStatsCardDoesNotContain('rows-card', EXPECTED_STATS.ROWS);
 
     // Columns card - should not show value when fields are empty
     await statsPage.verifyLatestStatsCardDoesNotContain('columns-card', '5');
